+++ src/main/java/com/shultzlab/weighttrackerapi/controllers/WeightEntryController.java
@@ -1,11 +1,13 @@
 package com.shultzlab.weighttrackerapi.controllers;
 
 import com.shultzlab.weighttrackerapi.exceptions.ResourceNotFoundException;
+import com.shultzlab.weighttrackerapi.exceptions.TokenForbiddenException;
 import com.shultzlab.weighttrackerapi.models.User;
 import com.shultzlab.weighttrackerapi.models.WeightEntry;
 import com.shultzlab.weighttrackerapi.models.requests.WeightEntryRequest;
 import com.shultzlab.weighttrackerapi.repositories.UserRepository;
 import com.shultzlab.weighttrackerapi.repositories.WeightEntryRepository;
+import com.shultzlab.weighttrackerapi.services.TokenService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -22,41 +24,47 @@ public class WeightEntryController {
     final WeightEntryRepository weightEntryRepository;
     final UserRepository userRepository;
 
-    public WeightEntryController(WeightEntryRepository weightEntryRepository, UserRepository userRepository){
+    public WeightEntryController(WeightEntryRepository weightEntryRepository, UserRepository userRepository) {
         this.weightEntryRepository = weightEntryRepository;
         this.userRepository = userRepository;
     }
 
-    // TODO Get rid of this after token?
-    @GetMapping()
-    public List<WeightEntry> getAllWeightEntries(){
-        return this.weightEntryRepository.findAll();
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity<WeightEntry> getWeightEntryById(@PathVariable(value = "id") Long entryId) throws ResourceNotFoundException {
-        WeightEntry entry = this.weightEntryRepository.findById(entryId).orElseThrow(() -> new ResourceNotFoundException("Weight Entry not found for this id: " + entryId));
-        // TODO verify record belongs to user before returning
-        return ResponseEntity.ok().body(entry);
-    }
-
-    @GetMapping("/user/{id}")
-    public List<WeightEntry> getAllWeightEntriesForUser(@PathVariable(value = "id") Long userId) {
-        // TODO verify userId matches token
-        return this.weightEntryRepository.findAllByUserId(userId);
-    }
+    // Commented out as not used by client
+//    @GetMapping()
+//    public List<WeightEntry> getAllWeightEntries(){
+//        return this.weightEntryRepository.findAll();
+//    }
+//
+//    @GetMapping("/{id}")
+//    public ResponseEntity<WeightEntry> getWeightEntryById(@PathVariable(value = "id") Long entryId) throws ResourceNotFoundException {
+//        WeightEntry entry = this.weightEntryRepository.findById(entryId).orElseThrow(() -> new ResourceNotFoundException("Weight Entry not found for this id: " + entryId));
+//        // TODO verify record belongs to user before returning
+//        return ResponseEntity.ok().body(entry);
+//    }
+//
+//    @GetMapping("/user/{id}")
+//    public List<WeightEntry> getAllWeightEntriesForUser(@PathVariable(value = "id") Long userId) {
+//        // TODO verify userId matches token
+//        return this.weightEntryRepository.findAllByUserId(userId);
+//    }
 
     @GetMapping("/username/{username}")
     public List<WeightEntry> getAllWeightEntriesForUserScopedByDays(@PathVariable(value = "username") String username,
-                                                                    @RequestParam Optional<String> time) {
+                                                                    @RequestParam Optional<String> time,
+                                                                    @RequestHeader("x-auth-token") String token) throws TokenForbiddenException {
+
+        String tokenUser = TokenService.getUsernameFromToken(token);
+        if (!username.equals(tokenUser)) {
+            throw new TokenForbiddenException();
+        }
 
         String daysStr = "-1";
-        if(time.isPresent()){
+        if (time.isPresent()) {
             daysStr = time.get();
         }
 
         // Account for empty string
-        if(daysStr.equals("")){
+        if (daysStr.equals("")) {
             daysStr = "-1";
         }
 
@@ -65,7 +73,7 @@ public class WeightEntryController {
         LocalDate date = LocalDate.now().minusDays(days);
 
         // Handle all
-        if(days == -1){
+        if (days == -1) {
             date = LocalDate.EPOCH;
         }
 
@@ -74,21 +82,31 @@ public class WeightEntryController {
     }
 
     @GetMapping("/username/{username}/last")
-    public WeightEntry getLastWeightEntryByUsername(@PathVariable(value = "username") String username) throws ResourceNotFoundException  {
-        // TODO verify username matches token
+    public WeightEntry getLastWeightEntryByUsername(@PathVariable(value = "username") String username, @RequestHeader("x-auth-token") String token) throws ResourceNotFoundException, TokenForbiddenException {
         User user = this.userRepository.findDistinctTopByUsername(username);
-        if(user == null){
+
+        String tokenUser = TokenService.getUsernameFromToken(token);
+        if (!user.getUsername().equals(tokenUser)) {
+            throw new TokenForbiddenException();
+        }
+
+        if (user == null) {
             throw new ResourceNotFoundException("User not found with username: " + username);
         }
         return this.weightEntryRepository.findDistinctFirstByUserOrderByEntryDateDesc(user);
     }
 
     @PostMapping()
-    public WeightEntry createWeightEntry(@RequestBody WeightEntryRequest entry) throws ResourceNotFoundException {
-        // TODO verify user matches token
+    public WeightEntry createWeightEntry(@RequestBody WeightEntryRequest entry, @RequestHeader("x-auth-token") String token) throws ResourceNotFoundException, TokenForbiddenException {
         String username = entry.getUsername();
+
+        String tokenUser = TokenService.getUsernameFromToken(token);
+        if (!username.equals(tokenUser)) {
+            throw new TokenForbiddenException();
+        }
+
         User user = this.userRepository.findDistinctTopByUsername(username);
-        if(user == null){
+        if (user == null) {
             throw new ResourceNotFoundException("User not found with username: " + username);
         }
         WeightEntry newWeightEntry = new WeightEntry(user, entry.getWeight(), entry.getEntryDate());
@@ -96,11 +114,16 @@ public class WeightEntryController {
     }
 
     @PutMapping("/{id}")
-    public WeightEntry updateWeightEntry(@PathVariable(value = "id") Long entryId, @RequestBody WeightEntryRequest entryRequest) throws ResourceNotFoundException {
-        // TODO verify user matches token
+    public WeightEntry updateWeightEntry(@PathVariable(value = "id") Long entryId, @RequestBody WeightEntryRequest entryRequest, @RequestHeader("x-auth-token") String token) throws ResourceNotFoundException, TokenForbiddenException {
         String username = entryRequest.getUsername();
+
+        String tokenUser = TokenService.getUsernameFromToken(token);
+        if (!username.equals(tokenUser)) {
+            throw new TokenForbiddenException();
+        }
+
         User user = this.userRepository.findDistinctTopByUsername(username);
-        if(user == null){
+        if (user == null) {
             throw new ResourceNotFoundException("User not found with username: " + username);
         }
         WeightEntry entry = this.weightEntryRepository.findById(entryId).orElseThrow(() -> new ResourceNotFoundException("Entry not found with id: " + entryId));
@@ -111,9 +134,15 @@ public class WeightEntryController {
     }
 
     @DeleteMapping("/{id}")
-    public Map<String, Boolean> deleteWeightEntry(@PathVariable(value = "id")Long entryId) throws ResourceNotFoundException {
+    public Map<String, Boolean> deleteWeightEntry(@PathVariable(value = "id") Long entryId, @RequestHeader("x-auth-token") String token) throws ResourceNotFoundException, TokenForbiddenException {
         WeightEntry entry = this.weightEntryRepository.findById(entryId).orElseThrow(() -> new ResourceNotFoundException("Entry not found with id: " + entryId));
-        // TODO verify record belongs to user before deleting
+
+        String tokenUser = TokenService.getUsernameFromToken(token);
+        if (!entry.getUser().getUsername().equals(tokenUser)) {
+            throw new TokenForbiddenException();
+        }
+
+
         this.weightEntryRepository.delete(entry);
         Map<String, Boolean> response = new HashMap<>();
         response.put("deleted", Boolean.TRUE);
